+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
-
+## 0.2 *(2019-09-29)*
+*   volume bar on album page
+*   rememeber volume
+*   show buffer state on discography page
 ## [0.1](https://github.com/cvzi/Bandcamp-script-deluxe-edition/tree/v0.1) *(2019-09-22)*
 *   player on discography page
 *   save albums as "played" by clicking on a "✔ Mark as played" link